fixed issue #112: Options panel misaligned on widescreen resolutions.
https://github.com/jinzhu/vrome/issues/112
The cause of misalignment was the combination of 'width: 900px;' and
'overflow: auto;' in css style of tag section.  A screen resolution with
1920x1080 may just be wide enough to put that <section> side by side
with <nav>, it depends on other factor such as the width of window
border of the browser, etc. I changed overflow to visible, which is the
default value.